Fix title to use Title object.
Apparently upgrade to Bokeh 2.4 changed the figure.title to a Title
object, so the title string now needs to be accessed using Title.text.

This is not easily evident from
https://docs.bokeh.org/en/latest/docs/releases.html or the 2.4
changelog: https://github.com/bokeh/bokeh/blob/2.4.0/CHANGELOG so it
may have been a fixed bug that our behavior was dependent on.